[REF] Core: handle insertText and applyFormat in Char plugin
