+++ packages/core/src/VDocument.ts
@@ -1,6 +1,6 @@
 import { VNode } from './VNodes/VNode';
 import { VRange } from './VRange';
-import { CharNode, FormatType, FORMAT_TYPES } from '../../plugin-char/CharNode';
+import { FormatType } from '../../plugin-char/CharNode';
 import { withMarkers } from '../../utils/src/markers';
 import { FragmentNode } from './VNodes/FragmentNode';
 import { VSelection } from './VSelection';
@@ -48,54 +48,6 @@ export class VDocument {
         range.start.before(node);
     }
     /**
-     * Insert text at the current position of the selection.
-     *
-     * If the selection is collapsed, add `text` to the vDocument and copy the
-     * formating of the previous char or the next char.
-     *
-     * If the selection is not collapsed, replace the text with the formating
-     * that was present in the selection.
-     *
-     * @param text
-     */
-    insertText(text: string, range = this.selection.range): void {
-        const format = this._getCurrentFormat();
-        // Remove the contents of the range if needed.
-        if (!range.isCollapsed()) {
-            this.deleteSelection(range);
-        }
-        // Split the text into CHAR nodes and insert them at the range.
-        const characters = text.split('');
-        characters.forEach(char => {
-            const vNode = new CharNode(char, format);
-            range.start.before(vNode);
-        });
-        this.formatCache = null;
-    }
-
-    /**
-     * Get the format for the next insertion.
-     */
-    _getCurrentFormat(range = this.selection.range): FormatType {
-        let format: FormatType = {};
-        if (this.formatCache) {
-            return this.formatCache;
-        } else if (range.isCollapsed()) {
-            const charToCopyFormat = range.start.previousSibling(CharNode) ||
-                range.start.nextSibling(CharNode) || {
-                    format: {},
-                };
-            format = { ...charToCopyFormat.format };
-        } else {
-            const selectedChars = range.selectedNodes(CharNode);
-            FORMAT_TYPES.forEach(formatName => {
-                format[formatName] = selectedChars.some(char => char.format[formatName]);
-            });
-        }
-        return format;
-    }
-
-    /**
      * Truncate the tree by removing the selected nodes and merging their
      * orphaned children into the parent of the first removed node.
      */
@@ -123,36 +75,6 @@ export class VDocument {
     // Context
     //--------------------------------------------------------------------------
 
-    /**
-     * Apply the `format` to the selection.
-     *
-     * If the selection is collapsed, set the format on the selection so we know
-     * in the next insert which format should be used.
-     */
-    applyFormat(formatName: string, range = this.selection.range): void {
-        if (range.isCollapsed()) {
-            if (!this.formatCache) {
-                this.formatCache = this._getCurrentFormat();
-            }
-            this.formatCache[formatName] = !this.formatCache[formatName];
-        } else {
-            const selectedChars = range.selectedNodes(CharNode);
-
-            // If there is no char with the format `formatName` in the
-            // selection, set the format to true for all nodes.
-            if (!selectedChars.every(char => char.format[formatName])) {
-                selectedChars.forEach(char => {
-                    char[formatName] = true;
-                });
-                // If there is at least one char in with the format `fomatName`
-                // in the selection, set the format to false for all nodes.
-            } else {
-                selectedChars.forEach(char => {
-                    char[formatName] = false;
-                });
-            }
-        }
-    }
     /**
      * Insert/remove a list at range.
      *
